[Add]ステータス機能連動
