+++ app/assets/stylesheets/admin/order_details.scss
@@ -1,0 +1,3 @@
+// Place all the styles related to the admin::order_details controller here.
+// They will automatically be included in application.css.
+// You can use Sass (SCSS) here: http://sass-lang.com/
