+++ app/helpers/admin/order_details_helper.rb
@@ -1,0 +1,2 @@
+module Admin::OrderDetailsHelper
+end
