+++ test/controllers/admin/order_details_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class Admin::OrderDetailsControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
